+++ APIV2.Mark.Database/Models/OrderItem.cs
@@ -7,7 +7,7 @@ namespace APIV2.Mark.Database.Models
     public class OrderItem
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int ProductId { get; set; }
         public double Quantity { get; set; }
         public double Price { get; set; }
@@ -16,7 +16,7 @@ namespace APIV2.Mark.Database.Models
         public double? Total { get; set; }
         public double? TotalBeforeVatAndDiscount { get; set; }
         public double? Discount { get; set; }
-        public int OrderId { get; set; }
+        public long OrderId { get; set; }
         public string? Notes { get; set; }
 
         public Order Order { get; set; } = null!;
