+++ APIV2.Mark.Database/Models/Payments.cs
@@ -10,13 +10,13 @@ namespace APIV2.Mark.Database.Models
     public class Payment
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public long AccountId { get; set; }
         public int PurchaseId { get; set; }
         public string Code { get; set; } = null!;
         public float Amount { get; set; }
-        public int VendorId { get; set; }
-        public int EmpId { get; set; }
+        public long VendorId { get; set; }
+        public long EmpId { get; set; }
         public int StatusId { get; set; }
         public DateTime CreationDate { get; set; } = DateTime.Now;
         public string? Notes { get; set; }
